Split out fulfillment error reporting a bit more
